add: arrow key navigation in t and z
